+++ blockchain/app/app.go
@@ -157,12 +157,13 @@ import (
 	policymodule "github.com/qredo/fusionchain/x/policy"
 	policymodulekeeper "github.com/qredo/fusionchain/x/policy/keeper"
 	policymoduletypes "github.com/qredo/fusionchain/x/policy/types"
-	qassetsmodule "github.com/qredo/fusionchain/x/qassets"
-	qassetsmodulekeeper "github.com/qredo/fusionchain/x/qassets/keeper"
-	qassetsmoduletypes "github.com/qredo/fusionchain/x/qassets/types"
-	"github.com/qredo/fusionchain/x/revenue/v1"
-	revenuekeeper "github.com/qredo/fusionchain/x/revenue/v1/keeper"
-	revenuetypes "github.com/qredo/fusionchain/x/revenue/v1/types"
+
+	// qassetsmodule "github.com/qredo/fusionchain/x/qassets"
+	// qassetsmodulekeeper "github.com/qredo/fusionchain/x/qassets/keeper"
+	// qassetsmoduletypes "github.com/qredo/fusionchain/x/qassets/types"
+	// "github.com/qredo/fusionchain/x/revenue/v1"
+	// revenuekeeper "github.com/qredo/fusionchain/x/revenue/v1/keeper"
+	// revenuetypes "github.com/qredo/fusionchain/x/revenue/v1/types"
 	treasurymodule "github.com/qredo/fusionchain/x/treasury"
 	treasurymodulekeeper "github.com/qredo/fusionchain/x/treasury/keeper"
 	treasurymoduletypes "github.com/qredo/fusionchain/x/treasury/types"
@@ -251,12 +252,12 @@ var (
 		// Fusion modules
 		evm.AppModuleBasic{},
 		feemarket.AppModuleBasic{},
-		revenue.AppModuleBasic{},
+		// revenue.AppModuleBasic{},
 		wasm.AppModuleBasic{},
 		policymodule.AppModuleBasic{},
 		identitymodule.AppModuleBasic{},
 		treasurymodule.AppModuleBasic{},
-		qassetsmodule.AppModuleBasic{},
+		// qassetsmodule.AppModuleBasic{},
 	)
 
 	// module account permissions
@@ -272,7 +273,7 @@ var (
 		govtypes.ModuleName:            {authtypes.Burner},
 		ibctransfertypes.ModuleName:    {authtypes.Minter, authtypes.Burner},
 		evmtypes.ModuleName:            {authtypes.Minter, authtypes.Burner}, // used for secure addition and subtraction of balance using module account
-		qassetsmoduletypes.ModuleName:  {authtypes.Minter, authtypes.Burner}, // qAsset minting/burning
+		// qassetsmoduletypes.ModuleName:  {authtypes.Minter, authtypes.Burner}, // qAsset minting/burning
 	}
 
 	// module accounts that are allowed to receive tokens
@@ -339,12 +340,12 @@ type FusionApp struct {
 	// Fusion keepers
 	EvmKeeper       *evmkeeper.Keeper
 	FeeMarketKeeper feemarketkeeper.Keeper
-	RevenueKeeper   revenuekeeper.Keeper
-	WasmKeeper      qredowasmkeeper.Keeper
-	IdentityKeeper  identitymodulekeeper.Keeper
-	TreasuryKeeper  treasurymodulekeeper.Keeper
-	PolicyKeeper    policymodulekeeper.Keeper
-	QAssetsKeeper   qassetsmodulekeeper.Keeper
+	// RevenueKeeper   revenuekeeper.Keeper
+	WasmKeeper     qredowasmkeeper.Keeper
+	IdentityKeeper identitymodulekeeper.Keeper
+	TreasuryKeeper treasurymodulekeeper.Keeper
+	PolicyKeeper   policymodulekeeper.Keeper
+	// QAssetsKeeper   qassetsmodulekeeper.Keeper
 
 	// the module manager
 	mm *module.Manager
@@ -429,12 +430,12 @@ func NewFusionApp(
 		// Fusion keys
 		evmtypes.StoreKey,
 		feemarkettypes.StoreKey,
-		revenuetypes.StoreKey,
+		// revenuetypes.StoreKey,
 		wasmtypes.StoreKey,
 		policymoduletypes.StoreKey,
 		identitymoduletypes.StoreKey,
 		treasurymoduletypes.StoreKey,
-		qassetsmoduletypes.StoreKey,
+		// qassetsmoduletypes.StoreKey,
 	)
 
 	// Add the EVM transient store key
@@ -643,7 +644,7 @@ func NewFusionApp(
 		&app.IBCKeeper.PortKeeper,
 		scopedWasmKeeper,
 		app.PolicyKeeper,
-		app.QAssetsKeeper,
+		// app.QAssetsKeeper,
 		app.TransferKeeper,
 		app.MsgServiceRouter(),
 		app.GRPCQueryRouter(),
@@ -683,16 +684,16 @@ func NewFusionApp(
 	)
 	treasuryModule := treasurymodule.NewAppModule(appCodec, app.TreasuryKeeper, app.AccountKeeper, app.BankKeeper)
 
-	app.QAssetsKeeper = *qassetsmodulekeeper.NewKeeper(
-		appCodec,
-		keys[qassetsmoduletypes.StoreKey],
-		keys[qassetsmoduletypes.MemStoreKey],
-		app.GetSubspace(qassetsmoduletypes.ModuleName),
-		app.BankKeeper,
-		app.TreasuryKeeper,
-		app.IdentityKeeper,
-	)
-	qassetsModule := qassetsmodule.NewAppModule(appCodec, app.QAssetsKeeper, app.AccountKeeper, app.BankKeeper)
+	// app.QAssetsKeeper = *qassetsmodulekeeper.NewKeeper(
+	// 	appCodec,
+	// 	keys[qassetsmoduletypes.StoreKey],
+	// 	keys[qassetsmoduletypes.MemStoreKey],
+	// 	app.GetSubspace(qassetsmoduletypes.ModuleName),
+	// 	app.BankKeeper,
+	// 	app.TreasuryKeeper,
+	// 	app.IdentityKeeper,
+	// )
+	// qassetsModule := qassetsmodule.NewAppModule(appCodec, app.QAssetsKeeper, app.AccountKeeper, app.BankKeeper)
 
 	// register the proposal types
 	govRouter := govv1beta1.NewRouter()
@@ -719,20 +720,20 @@ func NewFusionApp(
 		),
 	)
 
-	app.RevenueKeeper = revenuekeeper.NewKeeper(
-		keys[revenuetypes.StoreKey],
-		appCodec,
-		authtypes.NewModuleAddress(govtypes.ModuleName),
-		app.BankKeeper,
-		app.DistrKeeper,
-		app.AccountKeeper,
-		app.EvmKeeper,
-		authtypes.FeeCollectorName,
-	)
+	// app.RevenueKeeper = revenuekeeper.NewKeeper(
+	// 	keys[revenuetypes.StoreKey],
+	// 	appCodec,
+	// 	authtypes.NewModuleAddress(govtypes.ModuleName),
+	// 	app.BankKeeper,
+	// 	app.DistrKeeper,
+	// 	app.AccountKeeper,
+	// 	app.EvmKeeper,
+	// 	authtypes.FeeCollectorName,
+	// )
 
 	app.EvmKeeper = app.EvmKeeper.SetHooks(
 		evmkeeper.NewMultiEvmHooks(
-			app.RevenueKeeper.Hooks(),
+		// app.RevenueKeeper.Hooks(),
 		),
 	)
 
@@ -878,12 +879,12 @@ func NewFusionApp(
 		// Fusion app modules
 		evm.NewAppModule(app.EvmKeeper, app.AccountKeeper, evmSs),
 		feemarket.NewAppModule(app.FeeMarketKeeper, feeMarketSs),
-		revenue.NewAppModule(app.RevenueKeeper, app.AccountKeeper, app.GetSubspace(revenuetypes.ModuleName)),
+		// revenue.NewAppModule(app.RevenueKeeper, app.AccountKeeper, app.GetSubspace(revenuetypes.ModuleName)),
 		wasm.NewAppModule(appCodec, &app.WasmKeeper, app.StakingKeeper, app.AccountKeeper, app.BankKeeper, app.MsgServiceRouter(), app.GetSubspace(wasmtypes.ModuleName)),
 		identityModule,
 		treasuryModule,
 		policyModule,
-		qassetsModule,
+		// qassetsModule,
 	)
 
 	// During begin block slashing happens after distr.BeginBlocker so that
@@ -912,7 +913,7 @@ func NewFusionApp(
 		genutiltypes.ModuleName,
 		authz.ModuleName,
 		feegrant.ModuleName,
-		revenuetypes.ModuleName,
+		// revenuetypes.ModuleName,
 		nft.ModuleName,
 		group.ModuleName,
 		paramstypes.ModuleName,
@@ -924,7 +925,7 @@ func NewFusionApp(
 		policymoduletypes.ModuleName,
 		identitymoduletypes.ModuleName,
 		treasurymoduletypes.ModuleName,
-		qassetsmoduletypes.ModuleName,
+		// qassetsmoduletypes.ModuleName,
 	)
 
 	// NOTE: fee market module must go last in order to retrieve the block gas used.
@@ -947,7 +948,7 @@ func NewFusionApp(
 		evidencetypes.ModuleName,
 		authz.ModuleName,
 		feegrant.ModuleName,
-		revenuetypes.ModuleName,
+		// revenuetypes.ModuleName,
 		nft.ModuleName,
 		group.ModuleName,
 		paramstypes.ModuleName,
@@ -960,7 +961,7 @@ func NewFusionApp(
 		policymoduletypes.ModuleName,
 		identitymoduletypes.ModuleName,
 		treasurymoduletypes.ModuleName,
-		qassetsmoduletypes.ModuleName,
+		// qassetsmoduletypes.ModuleName,
 	)
 
 	// NOTE: The genutils module must occur after staking so that pools are
@@ -992,7 +993,7 @@ func NewFusionApp(
 		ibctransfertypes.ModuleName,
 		authz.ModuleName,
 		feegrant.ModuleName,
-		revenuetypes.ModuleName,
+		// revenuetypes.ModuleName,
 		nft.ModuleName,
 		group.ModuleName,
 		paramstypes.ModuleName,
@@ -1003,7 +1004,7 @@ func NewFusionApp(
 		policymoduletypes.ModuleName,
 		identitymoduletypes.ModuleName,
 		treasurymoduletypes.ModuleName,
-		qassetsmoduletypes.ModuleName,
+		// qassetsmoduletypes.ModuleName,
 		// NOTE: crisis module must go at the end to check for invariants on each module
 		crisistypes.ModuleName,
 	)
@@ -1291,7 +1292,7 @@ func initParamsKeeper(appCodec codec.BinaryCodec, legacyAmino *codec.LegacyAmino
 	// Fusion subspaces
 	paramsKeeper.Subspace(evmtypes.ModuleName).WithKeyTable(evmtypes.ParamKeyTable()) //nolint: staticcheck
 	paramsKeeper.Subspace(feemarkettypes.ModuleName).WithKeyTable(feemarkettypes.ParamKeyTable())
-	paramsKeeper.Subspace(revenuetypes.ModuleName)
+	// paramsKeeper.Subspace(revenuetypes.ModuleName)
 	paramsKeeper.Subspace(wasmtypes.ModuleName)
 	return paramsKeeper
 }