+++ blockchain/x/identity/keeper/msg_server_append_child_workspace_test.go
@@ -21,67 +21,81 @@ import (
 	"github.com/qredo/fusionchain/x/identity/types"
 )
 
-var defaultWs = types.Workspace{
-	Address:       "qredoworkspace14a2hpadpsy9h5m6us54",
+var childWs = types.Workspace{
+	Address:       "childWs",
 	Creator:       "testOwner",
 	Owners:        []string{"testOwner"},
 	AdminPolicyId: 0,
 	SignPolicyId:  0,
 }
 
-func Test_msgServer_AddWorkspaceOwner(t *testing.T) {
+var wsWithChild = types.Workspace{
+	Address:         "qredoworkspace14a2hpadpsy9h5m6us54",
+	Creator:         "testOwner",
+	Owners:          []string{"testOwner"},
+	ChildWorkspaces: []string{"childWs"},
+	AdminPolicyId:   0,
+	SignPolicyId:    0,
+}
+
+func Test_msgServer_AppendChildWorkspace(t *testing.T) {
 	type args struct {
 		workspace *types.Workspace
-		msg       *types.MsgAddWorkspaceOwner
+		childWs   *types.Workspace
+		msg       *types.MsgAppendChildWorkspace
 	}
 	tests := []struct {
 		name          string
 		args          args
-		want          *types.MsgAddWorkspaceOwnerResponse
+		want          *types.MsgAppendChildWorkspaceResponse
 		wantWorkspace *types.Workspace
 		wantErr       bool
 	}{
 		{
-			name: "PASS: add workspace owner",
+			name: "PASS: add child workspace",
 			args: args{
 				workspace: &defaultWs,
-				msg:       types.NewMsgAddWorkspaceOwner("testOwner", "qredoworkspace14a2hpadpsy9h5m6us54", "testOwner2", 100),
+				childWs:   &childWs,
+				msg:       types.NewMsgAppendChildWorkspace("testOwner", "qredoworkspace14a2hpadpsy9h5m6us54", childWs.Address, 100),
 			},
-			want: &types.MsgAddWorkspaceOwnerResponse{},
+			want: &types.MsgAppendChildWorkspaceResponse{},
 			wantWorkspace: &types.Workspace{
-				Address:       "qredoworkspace14a2hpadpsy9h5m6us54",
-				Creator:       "testOwner",
-				Owners:        []string{"testOwner", "testOwner2"},
-				AdminPolicyId: 0,
-				SignPolicyId:  0,
+				Address:         "qredoworkspace14a2hpadpsy9h5m6us54",
+				Creator:         "testOwner",
+				Owners:          []string{"testOwner"},
+				ChildWorkspaces: []string{"childWs"},
+				AdminPolicyId:   0,
+				SignPolicyId:    0,
 			},
-			wantErr: false,
 		},
 		{
 			name: "FAIL: workspace is nil or not found",
 			args: args{
 				workspace: &defaultWs,
-				msg:       types.NewMsgAddWorkspaceOwner("testOwner", "notAWorkspace", "testOwner2", 100),
+				childWs:   &childWs,
+				msg:       types.NewMsgAppendChildWorkspace("testOwner", "notAWorkspace", childWs.Address, 100),
 			},
-			want:    &types.MsgAddWorkspaceOwnerResponse{},
+			want:    &types.MsgAppendChildWorkspaceResponse{},
 			wantErr: true,
 		},
 		{
-			name: "FAIL: owner is already owner",
+			name: "FAIL: creator is not an owner",
 			args: args{
 				workspace: &defaultWs,
-				msg:       types.NewMsgAddWorkspaceOwner("testOwner", "qredoworkspace14a2hpadpsy9h5m6us54", "testOwner", 100),
+				childWs:   &childWs,
+				msg:       types.NewMsgAppendChildWorkspace("notAnOwner", "qredoworkspace14a2hpadpsy9h5m6us54", childWs.Address, 100),
 			},
-			want:    &types.MsgAddWorkspaceOwnerResponse{},
+			want:    &types.MsgAppendChildWorkspaceResponse{},
 			wantErr: true,
 		},
 		{
-			name: "FAIL: creator is no admin owner",
+			name: "FAIL: new child is already a child",
 			args: args{
-				workspace: &defaultWs,
-				msg:       types.NewMsgAddWorkspaceOwner("noOwner", "qredoworkspace14a2hpadpsy9h5m6us54", "testOwner", 100),
+				workspace: &wsWithChild,
+				childWs:   &childWs,
+				msg:       types.NewMsgAppendChildWorkspace("testOwner", "qredoworkspace14a2hpadpsy9h5m6us54", "childWs", 100),
 			},
-			want:    &types.MsgAddWorkspaceOwnerResponse{},
+			want:    &types.MsgAppendChildWorkspaceResponse{},
 			wantErr: true,
 		},
 	}
@@ -94,24 +108,24 @@ func Test_msgServer_AddWorkspaceOwner(t *testing.T) {
 			msgSer := keeper.NewMsgServerImpl(*ik)
 
 			genesis := types.GenesisState{
-				Workspaces: []types.Workspace{*tt.args.workspace},
+				Workspaces: []types.Workspace{*tt.args.workspace, *tt.args.childWs},
 			}
 			identity.InitGenesis(ctx, *ik, genesis)
 
-			got, err := msgSer.AddWorkspaceOwner(goCtx, tt.args.msg)
+			got, err := msgSer.AppendChildWorkspace(goCtx, tt.args.msg)
 			if (err != nil) != tt.wantErr {
-				t.Fatalf("AddWorkspaceOwner() error = %v, wantErr %v", err, tt.wantErr)
+				t.Fatalf("AppendChildWorkspace() error = %v, wantErr %v", err, tt.wantErr)
 			}
 
 			if !tt.wantErr {
 				if !reflect.DeepEqual(got, tt.want) {
-					t.Errorf("AddWorkspaceOwner() got = %v, want %v", got, tt.want)
+					t.Errorf("AppendChildWorkspace() got = %v, want %v", got, tt.want)
 				}
 
 				gotWorkspace := ik.GetWorkspace(ctx, tt.args.workspace.Address)
 
 				if !reflect.DeepEqual(gotWorkspace, tt.wantWorkspace) {
-					t.Errorf("AddWorkspaceOwner() got = %v, want %v", gotWorkspace, tt.wantWorkspace)
+					t.Errorf("AppendChildWorkspace() got = %v, want %v", gotWorkspace, tt.wantWorkspace)
 				}
 			}
 		})