+++ blockchain/x/identity/keeper/msg_server_new_keyring_test.go
@@ -32,7 +32,7 @@ func Test_msgServer_NewKeyring(t *testing.T) {
 		wantErr     bool
 	}{
 		{
-			name: "create a keyring",
+			name: "PASS: create a keyring",
 			args: args{
 				msg: types.NewMsgNewKeyring("testCreator", "testDescription", 0, 0, 0),
 			},
@@ -46,12 +46,13 @@ func Test_msgServer_NewKeyring(t *testing.T) {
 				Fees:        &types.KeyringFees{KeyReq: 0, SigReq: 0},
 				IsActive:    true,
 			},
-			wantErr: false,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ik, ctx := keepertest.IdentityKeeper(t)
+			keepers := keepertest.NewTest(t)
+			ik := keepers.IdentityKeeper
+			ctx := keepers.Ctx
 			goCtx := sdk.WrapSDKContext(ctx)
 			msgSer := keeper.NewMsgServerImpl(*ik)
 			got, err := msgSer.NewKeyring(goCtx, tt.args.msg)