+++ blockchain/x/identity/keeper/msg_server_remove_workspace_owner.go
@@ -29,6 +29,14 @@ func (k msgServer) RemoveWorkspaceOwner(goCtx context.Context, msg *types.MsgRem
 		return nil, fmt.Errorf("workspace not found")
 	}
 
+	if !ws.IsOwner(msg.Creator) {
+		return nil, fmt.Errorf("creator is not an owner of the workspace")
+	}
+
+	if !ws.IsOwner(msg.Owner) {
+		return nil, fmt.Errorf("new owner is not an owner of the workspace")
+	}
+
 	act, err := k.policyKeeper.AddAction(ctx, msg.Creator, msg, ws.AdminPolicyId, msg.Btl, nil)
 	if err != nil {
 		return nil, err