+++ blockchain/x/identity/keeper/query_workspace_by_address_test.go
@@ -34,7 +34,7 @@ func TestKeeper_WorkspaceByAddress(t *testing.T) {
 		wantErr       bool
 	}{
 		{
-			name: "happy path",
+			name: "PASS: happy path",
 			args: args{
 				req: &types.QueryWorkspaceByAddressRequest{
 					Address: "qredoworkspace14a2hpadpsy9h5m6us54",
@@ -51,10 +51,9 @@ func TestKeeper_WorkspaceByAddress(t *testing.T) {
 					SignPolicyId:    0,
 				},
 			},
-			wantErr: false,
 		},
 		{
-			name: "req is nil",
+			name: "FAIL: req is nil",
 			args: args{
 				req:          nil,
 				msgWorkspace: types.NewMsgNewWorkspace("testOwner", 0, 0),
@@ -63,7 +62,7 @@ func TestKeeper_WorkspaceByAddress(t *testing.T) {
 			wantErr: true,
 		},
 		{
-			name: "wrong workspace address",
+			name: "FAIL: wrong workspace address",
 			args: args{
 				req: &types.QueryWorkspaceByAddressRequest{
 					Address: "wrongAddress",
@@ -85,7 +84,9 @@ func TestKeeper_WorkspaceByAddress(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ik, ctx := keepertest.IdentityKeeper(t)
+			keepers := keepertest.NewTest(t)
+			ik := keepers.IdentityKeeper
+			ctx := keepers.Ctx
 			goCtx := sdk.WrapSDKContext(ctx)
 			msgSer := keeper.NewMsgServerImpl(*ik)
 			_, err := msgSer.NewWorkspace(goCtx, tt.args.msgWorkspace)