+++ blockchain/x/wasm/keeper/keeper_cgo.go
@@ -17,7 +17,7 @@ import (
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 
 	policy "github.com/qredo/fusionchain/x/policy/keeper"
-	qassets "github.com/qredo/fusionchain/x/qassets/keeper"
+	// qassets "github.com/qredo/fusionchain/x/qassets/keeper"
 )
 
 // NewKeeper creates a new contract Keeper instance
@@ -34,7 +34,7 @@ func NewKeeper(
 	portKeeper types.PortKeeper,
 	capabilityKeeper types.CapabilityKeeper,
 	policyKeeper policy.Keeper,
-	qassetsKeeper qassets.Keeper,
+	// qassetsKeeper qassets.Keeper,
 	portSource types.ICS20TransferPortSource,
 	router keeper.MessageRouter,
 	grpcRouter keeper.GRPCQueryRouter,
@@ -44,7 +44,7 @@ func NewKeeper(
 	authority string,
 	opts ...keeper.Option,
 ) Keeper {
-	messenger := NewDefaultMessageHandler(router, ics4Wrapper, channelKeeper, capabilityKeeper, bankKeeper, qassetsKeeper, cdc, portSource)
+	messenger := NewDefaultMessageHandler(router, ics4Wrapper, channelKeeper, capabilityKeeper, bankKeeper /*qassetsKeeper, */, cdc, portSource)
 	opts = append(opts, keeper.WithMessageHandler(messenger), keeper.WithQueryHandlerDecorator(func(old keeper.WasmVMQueryHandler) keeper.WasmVMQueryHandler {
 		queryPlugins, ok := old.(keeper.QueryPlugins)
 		if !ok {
@@ -53,8 +53,8 @@ func NewKeeper(
 		return DefaultQueryPlugins(queryPlugins, policyKeeper)
 	}))
 	return Keeper{
-		Keeper:        keeper.NewKeeper(cdc, storeKey, accountKeeper, bankKeeper, stakingKeeper, distrKeeper, ics4Wrapper, channelKeeper, portKeeper, capabilityKeeper, portSource, router, grpcRouter, homeDir, wasmConfig, availableCapabilities, authority, opts...),
-		policyKeeper:  policyKeeper,
-		qassetsKeeper: qassetsKeeper,
+		Keeper:       keeper.NewKeeper(cdc, storeKey, accountKeeper, bankKeeper, stakingKeeper, distrKeeper, ics4Wrapper, channelKeeper, portKeeper, capabilityKeeper, portSource, router, grpcRouter, homeDir, wasmConfig, availableCapabilities, authority, opts...),
+		policyKeeper: policyKeeper,
+		// qassetsKeeper: qassetsKeeper,
 	}
 }