+++ blockchain/CHANGELOG.md
@@ -1,3 +1,24 @@
+## fusiond [1.8.1](https://github.com/qredo/fusionchain/compare/fusiond@1.8.0...fusiond@1.8.1) (2024-01-25)
+
+
+### Bug Fixes
+
+* **blockchain:** mapToDeterministicSlice ([#235](https://github.com/qredo/fusionchain/issues/235)) ([558a4f9](https://github.com/qredo/fusionchain/commit/558a4f9fb9433e4c44396f46daece330d4c6c49b))
+
+# fusiond [1.8.0](https://github.com/qredo/fusionchain/compare/fusiond@1.7.1...fusiond@1.8.0) (2024-01-25)
+
+
+### Features
+
+* **blockchain:** Valuepolicy ([#234](https://github.com/qredo/fusionchain/issues/234)) ([f2add5e](https://github.com/qredo/fusionchain/commit/f2add5eb5b585a0f8ece6599934c2e9504fa744b))
+
+## fusiond [1.7.1](https://github.com/qredo/fusionchain/compare/fusiond@1.7.0...fusiond@1.7.1) (2024-01-25)
+
+
+### Bug Fixes
+
+* testutils dependencies ([#233](https://github.com/qredo/fusionchain/issues/233)) ([40c73a1](https://github.com/qredo/fusionchain/commit/40c73a16c911c8065a7208552f24dfbecfb99cfb))
+
 # fusiond [1.7.0](https://github.com/qredo/fusionchain/compare/fusiond@1.6.3...fusiond@1.7.0) (2024-01-19)